[refactor] gemini 리뷰 적용
